# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/ChangSZ/mall-go/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/authorized": {
            "get": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "调用方列表",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.authorized"
                ],
                "summary": "调用方列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "第几页",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页显示条数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "调用方key",
                        "name": "business_key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "调用方secret",
                        "name": "business_secret",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "调用方对接人",
                        "name": "business_developer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "备注",
                        "name": "remark",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authorized.listResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "新增调用方",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.authorized"
                ],
                "summary": "新增调用方",
                "parameters": [
                    {
                        "type": "string",
                        "description": "调用方key",
                        "name": "business_key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "调用方对接人",
                        "name": "business_developer",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "备注",
                        "name": "remark",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authorized.createResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/authorized_api": {
            "get": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "调用方接口地址列表",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.authorized"
                ],
                "summary": "调用方接口地址列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authorized.listAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "授权调用方接口地址",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.authorized"
                ],
                "summary": "授权调用方接口地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HashID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "请求方法",
                        "name": "method",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "请求地址",
                        "name": "api",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authorized.createAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "authorized.createAPIResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
        "authorized.createResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
...
        "authorized.listAPIData": {
            "type": "object",
            "properties": {
                "api": {
                    "description": "调用方对接人",
                    "type": "string"
                },
                "business_key": {
                    "description": "调用方key",
                    "type": "string"
                },
                "hash_id": {
                    "description": "hashID",
                    "type": "string"
                },
                "method": {
                    "description": "调用方secret",
                    "type": "string"
                }
            }
        },
        "authorized.listAPIResponse": {
            "type": "object",
            "properties": {
                "business_key": {
                    "description": "调用方key",
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/authorized.listAPIData"
                    }
                }
            }
        },
        "authorized.listData": {
            "type": "object",
            "properties": {
                "business_developer": {
                    "description": "调用方对接人",
                    "type": "string"
                },
                "business_key": {
                    "description": "调用方key",
                    "type": "string"
                },
                "business_secret": {
                    "description": "调用方secret",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "hashid": {
                    "description": "hashid",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "is_used": {
                    "description": "是否启用 1:是 -1:否",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "updated_user": {
                    "description": "更新人",
                    "type": "string"
                }
            }
        },
        "authorized.listResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/authorized.listData"
                    }
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "current_page": {
                            "type": "integer"
                        },
                        "per_page_count": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "code.Failure": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "业务返回数据"
                },
                "message": {
                    "description": "描述信息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "LoginToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}